Fixed unwanted behavior when closing windows
When closing the car edit window there was always the dialog asking if you want to save the changes, no matter if changes have even been made. It was also possible to save unallowed values when using this dialog for the driver and car edit window. Both windows have been fixed and overall improved